zmiana nazwy aplikacji (#22)
