+++ app/build.gradle
@@ -6,7 +6,7 @@ apply plugin: "de.mannodermaus.android-junit5"
 android {
     compileSdkVersion 29
     defaultConfig {
-        applicationId "pl.gov.anna"
+        applicationId "pl.gov.mc.protego"
         minSdkVersion 22
         targetSdkVersion 29
         versionCode System.env.CI_JOB_ID ? System.env.CI_JOB_ID.toInteger() : 1
