--- app/src/main/kotlin/pl/gov/anna/gcs/GoogleCloudStorage.kt
@@ -1,4 +1,0 @@
-package pl.gov.anna.gcs
-
-class GoogleCloudStorage {
-}
