--- app/src/main/kotlin/pl/gov/anna/information/AppInformation.kt
@@ -1,11 +1,0 @@
-package pl.gov.anna.information
-
-import pl.gov.anna.BuildConfig
-
-class AppInformation {
-    val api: Int
-        get() = BuildConfig.SUPPORTED_API_VERSION
-
-    val version: Int
-        get() = BuildConfig.VERSION_CODE
-}
