--- app/src/main/kotlin/pl/gov/anna/ui/main/MainActivityViewModel.kt
@@ -1,15 +1,0 @@
-package pl.gov.anna.ui.main
-
-import androidx.lifecycle.ViewModel
-import pl.gov.anna.file.FileManager
-import pl.gov.anna.gcs.GoogleCloudStorage
-
-class MainActivityViewModel(
-    val fileManager: FileManager,
-    val googleCloudStorage: GoogleCloudStorage
-) : ViewModel() {
-
-    fun sendSampleFileToGcs() {
-
-    }
-}
