+++ app/src/main/kotlin/pl/gov/mc/protego/gcs/GoogleCloudStorage.kt
@@ -1,0 +1,4 @@
+package pl.gov.mc.protego.gcs
+
+class GoogleCloudStorage {
+}
