+++ app/src/main/kotlin/pl/gov/mc/protego/ui/main/MainActivity.kt
@@ -1,0 +1,25 @@
+package pl.gov.mc.protego.ui.main
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import kotlinx.android.synthetic.main.activity_main.*
+import org.koin.android.ext.android.inject
+import pl.gov.mc.protego.R
+import pl.gov.mc.protego.information.Session
+import pl.gov.mc.protego.ui.registration.RegistrationActivity
+
+class MainActivity : AppCompatActivity() {
+
+    private val session: Session by inject()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        logout_button.setOnClickListener {
+            session.logout()
+            startActivity(Intent(this, RegistrationActivity::class.java))
+            finish()
+        }
+    }
+}
