+++ app/src/main/kotlin/pl/gov/mc/protego/ui/validator/MsisdnValidator.kt
@@ -1,0 +1,11 @@
+package pl.gov.mc.protego.ui.validator
+
+import java.lang.IllegalArgumentException
+
+
+class InvalidMsisdnException(message: String) : IllegalArgumentException(message)
+
+class MsisdnValidator {
+    fun validate(msisdn: String) = msisdn.length == 9
+    fun validateWithCountryCode(msisdn: String) = msisdn.length == 12
+}
