+++ app/src/test/kotlin/pl/gov/mc/protego/SampleSpekTest.kt
@@ -1,0 +1,43 @@
+package pl.gov.mc.protego
+
+import junit.framework.TestCase.assertEquals
+import junit.framework.TestCase.assertTrue
+import org.spekframework.spek2.Spek
+import org.spekframework.spek2.style.gherkin.Feature
+
+object SetFeature: Spek({
+    Feature("Set") {
+        val set by memoized { mutableSetOf<String>() }
+
+        Scenario("adding items") {
+            When("adding foo") {
+                set.add("foo")
+            }
+
+            Then("it should have a size of 1") {
+                assertEquals(1, set.size)
+            }
+
+            Then("it should contain foo") {
+                assertTrue(set.contains("foo"))
+            }
+        }
+
+        Scenario("getting the first item") {
+            val item = "foo"
+            Given("a non-empty set")  {
+                set.add(item)
+            }
+
+            lateinit var result: String
+
+            When("getting the first item") {
+                result = set.first()
+            }
+
+            Then("it should return the first item") {
+                assertEquals(item, result)
+            }
+        }
+    }
+})
